Refactor utils/comments to utils/reviewComments to avoid ambiguity
