+++ templates/vue/src/components/Sidebar/NodeReview/ReviewButtons.vue
@@ -15,7 +15,7 @@
 </template>
 
 <script>
-import * as Comment from "@/utils/comments"
+import * as Comment from "@/utils/reviewComments"
 import { nodeStatus } from "@/utils/constants"
 
 export default {
