+++ templates/vue/src/components/Sidebar/NodeReview/index.vue
@@ -40,7 +40,7 @@
 import { mapState, mapGetters, mapActions } from "vuex"
 
 import { nodeStatus } from "@/utils/constants"
-import * as Comment from "@/utils/comments"
+import * as Comment from "@/utils/reviewComments"
 import * as wp from "@/services/wp"
 
 import ReviewLog from "./ReviewLog"
