+++ templates/vue/src/components/modals/NodeModal/index.vue
@@ -272,7 +272,7 @@ import PermissionsTable from "../common/PermissionsTable"
 import DeleteNodeButton from "./DeleteNodeButton"
 import { names } from "@/config/routes"
 import Helpers from "@/utils/Helpers"
-import * as Comment from "@/utils/comments"
+import * as Comment from "@/utils/reviewComments"
 import { sizes, nodeStatus } from "@/utils/constants"
 import { getLinkMetadata } from "@/services/LinkPreviewApi"
 import DragSelectModular from "@/utils/dragSelectModular"
